refactor(web/layout): relocate Sugiyama module and update imports; add shim for back-compat
