+++ web/modules/sugiyama-layout.d.ts
@@ -1,0 +1,1 @@
+export * from './layout/sugiyama-layout.js';
